modernizer/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
distro==1.9.0
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
openai==1.14.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zope.interface==6.2
